backend/requirements.txt
fastapi
uvicorn[standard]
gql[aiohttp]
websockets
pydantic
python-dotenv
